+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Basic CLI interface
 - Generation of data based on regexes and format properties
+- Support unique element generation in arrays
 
 ### Changed
 - Updated jsonoid-discovery dependency to v0.5.4